# admin-frontend/src/api/generated/models/AssistantChatRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type AssistantChatRequest = {
    groupId?: number;
    message?: string;
    topK?: number;
    templateId?: string;
    conversationId?: string;
};