# frontend/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/user/register":{"post":{"tags":["user-controller"],"operationId":"userRegister","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseLong"}}}}}}},"/user/logout":{"post":{"tags":["user-controller"],"operationId":"userLogout","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseBoolean"}}}}}}},"/user/login":{"post":{"tags":["user-controller"],"operationId":"userLogin","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseLoginUserVO"}}}}}}},"/rag/qa/ask":{"post":{"tags":["rag-controller"],"operationId":"ask","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QaAskRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseQaAnswerVO"}}}}}}},"/rag/ingest/text":{"post":{"tags":["rag-controller"],"operationId":"ingestText","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IngestTextRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseIngestionTaskVO"}}}}}}},"/rag/ingest/jobs/{jobId}/retry":{"post":{"tags":["rag-controller"],"operationId":"retryJob","parameters":[{"name":"jobId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseBoolean"}}}}}}},"/rag/ingest/file":{"post":{"tags":["rag-controller"],"operationId":"ingestFile","parameters":[{"name":"groupId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"required":["file"],"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseIngestionTaskVO"}}}}}}},"/rag/ingest/documents/{documentId}/rebuild":{"post":{"tags":["rag-controller"],"operationId":"rebuildDocument","parameters":[{"name":"documentId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseBoolean"}}}}}}},"/group/join":{"post":{"tags":["group-controller"],"operationId":"join","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JoinGroupRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseBoolean"}}}}}}},"/group/create":{"post":{"tags":["group-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateGroupRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseLong"}}}}}}},"/assistant/eval/complaint":{"post":{"tags":["assistant-controller"],"operationId":"evaluateComplaint","parameters":[{"name":"groupId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"templateId","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseMapStringObject"}}}}}}},"/assistant/chat":{"post":{"tags":["assistant-controller"],"operationId":"chat","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssistantChatRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/x-ndjson":{"schema":{"type":"array","items":{"type":"string"}}}}}}}},"/user/list":{"get":{"tags":["user-controller"],"operationId":"listUsers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseListLoginUserVO"}}}}}}},"/rag/ingest/task/{jobId}":{"get":{"tags":["rag-controller"],"operationId":"getTaskStatus","parameters":[{"name":"jobId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseIngestionTaskVO"}}}}}}},"/rag/ingest/metrics":{"get":{"tags":["rag-controller"],"operationId":"ingestionMetrics","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseIngestionMetricsVO"}}}}}}},"/rag/ingest/jobs":{"get":{"tags":["rag-controller"],"operationId":"listJobs","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseListIngestionJobVO"}}}}}}},"/rag/ingest/jobs/{jobId}":{"get":{"tags":["rag-controller"],"operationId":"getJob","parameters":[{"name":"jobId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseIngestionJobVO"}}}}}}},"/group/my/list":{"get":{"tags":["group-controller"],"operationId":"myGroups","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseListGroupVO"}}}}}}}},"components":{"schemas":{"UserRegisterRequest":{"type":"object","properties":{"userCode":{"type":"string"},"displayName":{"type":"string"},"userPassword":{"type":"string"},"checkPassword":{"type":"string"}}},"BaseResponseLong":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"type":"integer","format":"int64"},"message":{"type":"string"}}},"BaseResponseBoolean":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"type":"boolean"},"message":{"type":"string"}}},"UserLoginRequest":{"type":"object","properties":{"userCode":{"type":"string"},"userPassword":{"type":"string"}}},"BaseResponseLoginUserVO":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"$ref":"#/components/schemas/LoginUserVO"},"message":{"type":"string"}}},"LoginUserVO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"userCode":{"type":"string"},"displayName":{"type":"string"},"userRole":{"type":"string"},"token":{"type":"string"}}},"QaAskRequest":{"type":"object","properties":{"groupId":{"type":"integer","format":"int64"},"question":{"type":"string"},"topK":{"type":"integer","format":"int32"}}},"BaseResponseQaAnswerVO":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"$ref":"#/components/schemas/QaAnswerVO"},"message":{"type":"string"}}},"ChunkEvidenceVO":{"type":"object","properties":{"documentId":{"type":"integer","format":"int64"},"chunkIndex":{"type":"integer","format":"int32"},"fileName":{"type":"string"},"content":{"type":"string"},"charStart":{"type":"integer","format":"int32"},"charEnd":{"type":"integer","format":"int32"},"score":{"type":"number","format":"double"},"source":{"type":"string"}}},"QaAnswerVO":{"type":"object","properties":{"answer":{"type":"string"},"evidenceEnough":{"type":"boolean"},"confidenceLevel":{"type":"string"},"confidenceScore":{"type":"number","format":"double"},"evidences":{"type":"array","items":{"$ref":"#/components/schemas/ChunkEvidenceVO"}}}},"IngestTextRequest":{"type":"object","properties":{"groupId":{"type":"integer","format":"int64"},"fileName":{"type":"string"},"content":{"type":"string"}}},"BaseResponseIngestionTaskVO":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"$ref":"#/components/schemas/IngestionTaskVO"},"message":{"type":"string"}}},"IngestionTaskVO":{"type":"object","properties":{"documentId":{"type":"integer","format":"int64"},"jobId":{"type":"integer","format":"int64"},"documentStatus":{"type":"string"},"jobStatus":{"type":"string"},"failureReason":{"type":"string"},"lastError":{"type":"string"}}},"JoinGroupRequest":{"type":"object","properties":{"groupId":{"type":"integer","format":"int64"}}},"CreateGroupRequest":{"type":"object","properties":{"groupCode":{"type":"string"},"groupName":{"type":"string"}}},"BaseResponseMapStringObject":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"type":"object","additionalProperties":{"type":"object"}},"message":{"type":"string"}}},"AssistantChatRequest":{"type":"object","properties":{"groupId":{"type":"integer","format":"int64"},"message":{"type":"string"},"topK":{"type":"integer","format":"int32"},"templateId":{"type":"string"},"conversationId":{"type":"string"}}},"BaseResponseListLoginUserVO":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"type":"array","items":{"$ref":"#/components/schemas/LoginUserVO"}},"message":{"type":"string"}}},"BaseResponseIngestionMetricsVO":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"$ref":"#/components/schemas/IngestionMetricsVO"},"message":{"type":"string"}}},"IngestionMetricsVO":{"type":"object","properties":{"totalJobs":{"type":"integer","format":"int64"},"successJobs":{"type":"integer","format":"int64"},"failedJobs":{"type":"integer","format":"int64"},"failureRate":{"type":"number","format":"double"},"avgDurationSeconds":{"type":"number","format":"double"},"successRateByFileType":{"type":"object","additionalProperties":{"type":"number","format":"double"}},"retryCountDistribution":{"type":"object","additionalProperties":{"type":"integer","format":"int64"}}}},"BaseResponseListIngestionJobVO":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"type":"array","items":{"$ref":"#/components/schemas/IngestionJobVO"}},"message":{"type":"string"}}},"IngestionJobVO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"documentId":{"type":"integer","format":"int64"},"groupId":{"type":"integer","format":"int64"},"jobType":{"type":"string"},"status":{"type":"string"},"retryCount":{"type":"integer","format":"int32"},"maxRetries":{"type":"integer","format":"int32"},"startedAt":{"type":"string","format":"date-time"},"finishedAt":{"type":"string","format":"date-time"},"nextRetryAt":{"type":"string","format":"date-time"},"lastError":{"type":"string"}}},"BaseResponseIngestionJobVO":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"$ref":"#/components/schemas/IngestionJobVO"},"message":{"type":"string"}}},"BaseResponseListGroupVO":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"type":"array","items":{"$ref":"#/components/schemas/GroupVO"}},"message":{"type":"string"}}},"GroupVO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"groupCode":{"type":"string"},"groupName":{"type":"string"},"status":{"type":"string"}}}}}}